Add a root endpoint that returns more info
