+++ api/routes/root.py
@@ -1,5 +1,6 @@
 from flask import jsonify
 from flask import Blueprint
+from flask import request
 from flask_cors import CORS
 from flask_cors import cross_origin
 
@@ -7,10 +8,22 @@ mod = Blueprint('root', __name__)
 cors = CORS(mod)
 
 
-@mod.route("/")
-@cross_origin()
-def root():
+# This is some of the information returned at the root endpoint
+# The root endpoint is established in ../__init__.py
+def root_data():
     data = {
         "data": "Welcome to the Throne API."
     }
-    return jsonify(data)
+    return data
+
+
+@mod.route("/debug")
+@cross_origin()
+def debug():
+    my_list = [
+        str(request.headers),
+        str(request.url),
+        str(request.environ.get("serverless.context", "no context")),
+        str(request.environ.get("serverless.event", "no event")),
+    ]
+    return jsonify(my_list)
